requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-WTF==1.2.1
Werkzeug==3.0.1
pandas==2.1.4
openpyxl==3.1.2
python-dateutil==2.8.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
anthropic>=0.88.0
fpdf2>=2.8.0
